+++ src/__stories__/useGeolocation.story.tsx
@@ -13,7 +13,7 @@ const Demo = () => {
   );
 };
 
-storiesOf('Sensors/useGeolocation', module)
+storiesOf('Sensors|useGeolocation', module)
   .add('Docs', () => <ShowDocs md={require('../../docs/useGeolocation.md')} />)
   .add('Demo', () =>
     <Demo/>