+++ src/__stories__/useMount.story.tsx
@@ -8,7 +8,7 @@ const Demo = () => {
   return null;
 };
 
-storiesOf('Lifecycles/useMount', module)
+storiesOf('Lifecycles|useMount', module)
   .add('Docs', () => <ShowDocs md={require('../../docs/useMount.md')} />)
   .add('Demo', () =>
     <Demo/>