+++ src/__stories__/useObservable.story.tsx
@@ -7,7 +7,7 @@ import {BehaviorSubject} from 'rxjs';
 const counter$ = new BehaviorSubject(0);
 const Demo = () => {
   const value = useObservable(counter$, 0);
-  
+
   return (
     <button onClick={() => counter$.next(value! + 1)}>
       Clicked {value} times
@@ -15,8 +15,8 @@ const Demo = () => {
   );
 };
 
-storiesOf('State/useObservable', module)
-  // .add('Docs', () => <ShowDocs md={require('../../docs/useObservable.md')} />)
+storiesOf('State|useObservable', module)
+  .add('Docs', () => <ShowDocs md={require('../../docs/useObservable.md')} />)
   .add('Demo', () =>
     <Demo/>
   )