+++ src/__stories__/useOutsideClick.story.tsx
@@ -19,7 +19,7 @@ const Demo = () => {
   );
 };
 
-storiesOf('UI/useOutsideClick', module)
+storiesOf('UI|useOutsideClick', module)
   .add('Docs', () => <ShowDocs md={require('../../docs/useOutsideClick.md')} />)
   .add('Demo', () =>
     <Demo/>