+++ src/__stories__/useToggle.story.tsx
@@ -16,7 +16,7 @@ const Demo = () => {
   );
 };
 
-storiesOf('State/useToggle', module)
+storiesOf('State|useToggle', module)
   .add('Docs', () => <ShowDocs md={require('../../docs/useToggle.md')} />)
   .add('Demo', () =>
     <Demo/>