+++ src/__stories__/useWait.story.tsx
@@ -32,7 +32,7 @@ const Demo = () => {
   );
 };
 
-storiesOf('UI/useWait', module)
+storiesOf('UI|useWait', module)
   .add('Docs', () => <ShowDocs md={require('../../docs/useWait.md')} />)
   .add('Demo', () =>
     <useWait.Waiter><Demo/></useWait.Waiter>